% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{TicagrelorVsClopidogrelMiniReplica}
\alias{TicagrelorVsClopidogrelMiniReplica}
\alias{TicagrelorVsClopidogrelMiniReplica-package}
\title{TicagrelorVsClopidogrelMiniReplica}
\description{
TicagrelorVsClopidogrelMiniReplica
}
